backend/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
pandas==2.1.0
pydantic==1.10.12
python-dotenv==1.0.0